Unify proof creation by using self-MPC for single-prover.
This eliminates the `SinglePartyAggregator` and just makes it into a method on
the `RangeProof` type.
